firmware: Added transport module and added usbd
