+++ firmware/nrf/transport.h
@@ -1,0 +1,68 @@
+/*******************************************************************************
+ *
+ *      Copyright (c) 2019, Raccon BLE Sniffer
+ *      All rights reserved.
+ *
+ *      Redistribution and use in source and binary forms, with or without
+ *      modification, are permitted provided that the following conditions are
+ *      met:
+ *      
+ *      * Redistributions of source code must retain the above copyright
+ *        notice, this list of conditions and the following disclaimer.
+ *      * Redistributions in binary form must reproduce the above
+ *        copyright notice, this list of conditions and the following disclaimer
+ *        in the documentation and/or other materials provided with the
+ *        distribution.
+ *      * Neither the name of "btlejack2" nor the names of its
+ *        contributors may be used to endorse or promote products derived from
+ *        this software without specific prior written permission.
+ *      
+ *      THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ *      "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ *      LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ *      A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ *      OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ *      SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ *      LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ *      DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ *      THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ *      (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ *      OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ *******************************************************************************/
+
+/**
+ * transport
+ *
+ * This module provides all the required functions to manage the transport
+ * layer to host pc
+ **/
+
+#ifndef __TRANSPORT_H
+#define __TRANSPORT_H
+
+#include <stdint.h>
+#include <nrfx.h>
+
+/**
+ * @brief initialize transport layer
+ **/
+nrfx_err_t transport_init(void);
+
+/**
+ * @brief inform if transport has data to receive
+ **/
+bool transport_isready (void);
+
+/**
+ * @brief read transport data in blocking mode
+ **/
+nrfx_err_t transport_read(void* pbuf, size_t len);
+
+/**
+ * @brief write transport data in blocking mode
+ **/
+nrfx_err_t transport_write(void* pbuf, size_t len);
+
+
+#endif
